pyproject.toml
[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = ""
authors = [
    {name = "Serhii Hrekov",email = "themebeat@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pydantic (>=2.12.4,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
